CQTByMagnitudeDetector added + More tidying up

+++ MusicAnalyser/App/DSP/Scripts/CQTProcessor.cs
@@ -73,7 +73,7 @@ class CQTProcessor : ISignalProcessor
         }
 
         OutputBuffer = mag;
-        Func<int, double> scale = i => double.Parse(Settings["MIN_FREQ"][0]) * Math.Pow(2, i / int.Parse(Settings["BINS_PER_OCTAVE"][0]));
+        Func<int, double> scale = i => double.Parse(Settings["MIN_FREQ"][0]) * Math.Pow(2, (double)i / int.Parse(Settings["BINS_PER_OCTAVE"][0]));
         OutputScale = scale;
     }
